+++ src/client/terminals/envCollectionActivation/indicatorPrompt.ts
@@ -14,15 +14,15 @@ import {
 } from '../../common/types';
 import { Common, Interpreters } from '../../common/utils/localize';
 import { IExtensionSingleActivationService } from '../../activation/types';
-import { ITerminalEnvVarCollectionService } from './types';
+import { ITerminalEnvVarCollectionService } from '../../interpreter/activation/types';
 import { inTerminalEnvVarExperiment } from '../../common/experiments/helpers';
-import { IInterpreterService } from '../contracts';
+import { IInterpreterService } from '../../interpreter/contracts';
 import { PythonEnvironment } from '../../pythonEnvironments/info';
 
 export const terminalEnvCollectionPromptKey = 'TERMINAL_ENV_COLLECTION_PROMPT_KEY';
 
 @injectable()
-export class TerminalEnvVarCollectionPrompt implements IExtensionSingleActivationService {
+export class TerminalIndicatorPrompt implements IExtensionSingleActivationService {
     public readonly supportedWorkspaceTypes = { untrustedWorkspace: false, virtualWorkspace: false };
 
     constructor(
