+++ src/client/terminals/serviceRegistry.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { interfaces } from 'inversify';
-import { ClassType } from '../ioc/types';
+import { IServiceManager } from '../ioc/types';
 import { TerminalAutoActivation } from './activation';
 import { CodeExecutionManager } from './codeExecution/codeExecutionManager';
 import { DjangoShellCodeExecutionProvider } from './codeExecution/djangoShellCodeExecution';
@@ -10,16 +9,12 @@ import { CodeExecutionHelper } from './codeExecution/helper';
 import { ReplProvider } from './codeExecution/repl';
 import { TerminalCodeExecutionProvider } from './codeExecution/terminalCodeExecution';
 import { ICodeExecutionHelper, ICodeExecutionManager, ICodeExecutionService, ITerminalAutoActivation } from './types';
+import { IExtensionActivationService, IExtensionSingleActivationService } from '../activation/types';
+import { ITerminalEnvVarCollectionService } from '../interpreter/activation/types';
+import { TerminalIndicatorPrompt } from './envCollectionActivation/indicatorPrompt';
+import { TerminalEnvVarCollectionService } from './envCollectionActivation/service';
 
-interface IServiceRegistry {
-    addSingleton<T>(
-        serviceIdentifier: interfaces.ServiceIdentifier<T>,
-        constructor: ClassType<T>,
-        name?: string | number | symbol,
-    ): void;
-}
-
-export function registerTypes(serviceManager: IServiceRegistry): void {
+export function registerTypes(serviceManager: IServiceManager): void {
     serviceManager.addSingleton<ICodeExecutionHelper>(ICodeExecutionHelper, CodeExecutionHelper);
 
     serviceManager.addSingleton<ICodeExecutionManager>(ICodeExecutionManager, CodeExecutionManager);
@@ -37,4 +32,13 @@ export function registerTypes(serviceManager: IServiceRegistry): void {
     serviceManager.addSingleton<ICodeExecutionService>(ICodeExecutionService, ReplProvider, 'repl');
 
     serviceManager.addSingleton<ITerminalAutoActivation>(ITerminalAutoActivation, TerminalAutoActivation);
+    serviceManager.addSingleton<ITerminalEnvVarCollectionService>(
+        ITerminalEnvVarCollectionService,
+        TerminalEnvVarCollectionService,
+    );
+    serviceManager.addBinding(ITerminalEnvVarCollectionService, IExtensionActivationService);
+    serviceManager.addSingleton<IExtensionSingleActivationService>(
+        IExtensionSingleActivationService,
+        TerminalIndicatorPrompt,
+    );
 }
